Update Data.js
Some wording changes. Also added some (but not all) examples of the use of Element which has been used as a capitalised Thing in Stage 1. That's entirely a preference thing, but either way likely needs to be checked more thoroughly for consistency. Golem has been capitalised for similar reasons, but more consistently. Similarly with Stash upgrades, 'nth' may not be something you wish to include.